Added middle lane trajectory deque
Curve trajectory is now independent of side-to-side movement, so there
won't be any errors related to that. This is needed because moving
sode-to-side too fast previously could, in edge cases, lead to waypoints
being picked up non-sequentially, leading to the spline library
complaining.

All calculations are done with the middle_line_trajectory, not the
previous_path. Then, once all calculations are done, the actual position
is sent to the program, while the middle position is stored.