fix for while loop condition in getXY
the second part of the while condition needs to be fixed. otherwise we can get ``prev_wp == maps_s.size()-1`` which leads to an invalid memory access at the next check of the first part of the while condition ``s > map_s[prev_wp+1]``

This error only occurs for x,y positions beyond the last waypoint.