Fix computing angle to the closest waypoint
Angle must be normalized for the correct waypoint classification. Before
this change, if the car angle equals 3 and the waypoint angle equals -3,
the difference between them becomes 6. However, the actual difference
must be equal to 0.28.
After this change, the difference lies in the range [0, pi]. A
difference less than pi/2 means a forward waypoint. A difference in
the range [pi/2, pi] means a backward waypoint.